requirements.txt
chalice
httpie
pytest 
uplink
pytest-chalice
pytest-cov
pytest-xdist
mysql-connector-python